requirements.txt
Django==1.4.1
Fabric==1.3.3
simplejson==2.2.1
django-tokenapi
